+++ src/lib.rs
@@ -104,7 +104,8 @@ let kexp = cte::KEXP_DEFAULT; // factor de exportación [-]
 let arearef = 200.0; // superficie de referencia [m2]
 
 // Cálculo del balance global según EN ISO 52000-1:2017
-let ep = energy_performance(&components, &fp, kexp, arearef).unwrap();
+let load_matching = false;
+let ep = energy_performance(&components, &fp, kexp, arearef, load_matching).unwrap();
 
 // Visualización compacta
 println!("{}", &ep.to_plain());