+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Este proyecto sigue, además, el [Versionado semántico](https://semver.org/spec
   - Cambio de nombre de sufijos "_bygen" a "_by_src" y "_byuse" a "_by_srv" en la salida JSON
 - Ejecutable cteepbd:
   - Eliminada la opción --acsnrb para el cálculo exclusivo de ACS en perímetro nearby (ya se calcula incondicionalmente)
+  - Añadida la opción --load_matching para realizar el cálculo de coincidencia de cargas (en lugar de usar f_match = 1)
 
 ### Incompatibilidades